# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Real-time multi-broker MQTT sensor dashboard with JWT authentication. Provides device management, sensor readings with TimescaleDB downsampling, and field rename configuration.",
        "title": "TelemetryHub API",
        "termsOfService": "https://telemetryhub.local/terms",
        "contact": {
            "name": "API Support",
            "email": "admin@telemetryhub.local"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/devices/{id}/renames": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all field rename configurations for a device.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "renames"
                ],
                "summary": "List field renames",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_handlers.FieldRename"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/telemetryhub_internal_auth.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/telemetryhub_internal_auth.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new field rename configuration for a device. The display_name and unit are used in chart labels. Requires admin role.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "renames"
                ],
                "summary": "Create field rename",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Rename configuration",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.CreateRenameRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/telemetryhub_internal_auth.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/telemetryhub_internal_auth.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/telemetryhub_internal_auth.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/telemetryhub_internal_auth.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/telemetryhub_internal_auth.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/devices/{id}/renames/{field}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates an existing field rename configuration. Requires admin role.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "renames"
                ],
                "summary": "Update field rename",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Raw field name",
                        "name": "field",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated rename configuration",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.UpdateRenameRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/telemetryhub_internal_auth.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/telemetryhub_internal_auth.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/telemetryhub_internal_auth.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/telemetryhub_internal_auth.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/telemetryhub_internal_auth.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a field rename configuration. Readings will fall back to raw field names. Requires admin role.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "renames"
                ],
                "summary": "Delete field rename",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Raw field name",
                        "name": "field",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/telemetryhub_internal_auth.MessageResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/telemetryhub_internal_auth.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/telemetryhub_internal_auth.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/telemetryhub_internal_auth.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_handlers.CreateRenameRequest": {
            "type": "object",
            "properties": {
                "chart_group": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "raw_field": {
                    "type": "string"
                },
                "unit": {
                    "type": "string"
                }
            }
        },
...
        "internal_handlers.FieldRename": {
            "type": "object",
            "properties": {
                "chart_group": {
                    "type": "string"
                },
                "device_id": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "raw_field": {
                    "type": "string"
                },
                "unit": {
                    "type": "string"
                }
            }
        },
...
        "internal_handlers.UpdateRenameRequest": {
            "type": "object",
            "properties": {
                "chart_group": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "unit": {
                    "type": "string"
                }
            }
        },
        "telemetryhub_internal_auth.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "telemetryhub_internal_auth.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Bearer token obtained from /auth/login",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}